+++ src/main/java/com/plog/plogbackend/domain/post/service/PostService.java
@@ -3,8 +3,8 @@ package com.plog.plogbackend.domain.post.service;
 import com.plog.plogbackend.domain.Member.Member;
 import com.plog.plogbackend.domain.Member.repository.MemberRepository;
 import com.plog.plogbackend.domain.badge.event.BadgeGrantEvent;
-import com.plog.plogbackend.domain.place.PlaceRepository;
 import com.plog.plogbackend.domain.place.entity.Place;
+import com.plog.plogbackend.domain.place.repository.PlaceRepository;
 import com.plog.plogbackend.domain.post.controller.dto.request.post.TimePickerRequest;
 import com.plog.plogbackend.domain.post.controller.dto.response.PostTextResponse;
 import com.plog.plogbackend.domain.post.entity.PlaceCategory;
